Backend/requirements.txt
flask==2.2.5
ultralytics==8.2.60
pillow==10.2.0
opencv-python==4.8.1.78
flask-cors==4.0.0